project/dashboard/requirements.txt
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-WTF==0.14.3
google-cloud-bigquery==1.15.0  ## for global utils
google-cloud-storage==1.16.1   ## for global utils
oauthlib==3.0.1
psycopg2==2.8.5
PyYAML==5.3.1
redis>=3
requests==2.24.0
rq>=0.13
rq-dashboard
sentry_sdk==0.17.8
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
statsd==3.3.0
